+++ sys/install_miniconda.sh
@@ -2,15 +2,15 @@
 set -ex
 
 ## Installing Anaconda
-if [ -d "sys/soft/minconda2" ] 
+if [ -d "sys/soft/minconda3" ] 
 then
-    rm -r soft/miniconda2
+    rm -r soft/miniconda3
 else
     mkdir -p soft
 fi
 sysPath=$PWD
 cd /tmp
-wget https://repo.anaconda.com/miniconda/Miniconda2-latest-Linux-x86_64.sh -O miniconda.sh
-bash ./miniconda.sh -p $sysPath/soft/miniconda2
+wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
+bash ./miniconda.sh -b -p $sysPath/soft/miniconda3