+++ sys/set_bashrc_env.sh
@@ -11,7 +11,7 @@ printf "\n## PySeg system "$date" \n" >> soft/bashrc_pyseg_sys
 printf "\n## DisPerSe \n" >> soft/bashrc_pyseg_sys
 printf "export PATH="$PWD"/soft/disperse/0.9.24_pyseg_gcc7/build/bin:""$""PATH" >> soft/bashrc_pyseg_sys
 printf "\n## Conda python \n" >> soft/bashrc_pyseg_sys
-printf "export PATH="$PWD"/soft/miniconda2/bin:""$""PATH" >> soft/bashrc_pyseg_sys
+printf "export PATH="$PWD"/soft/miniconda3/bin:""$""PATH" >> soft/bashrc_pyseg_sys
 printf "\n# PySeg python code \n" >> soft/bashrc_pyseg_sys
 printf "export PYTHONPATH="$PWD"/../code:""$""PYTHONPATH" >> soft/bashrc_pyseg_sys
 printf "\n" >> soft/bashrc_pyseg_sys